+++ storybook/stories/wallets/send/WalletSend.stories.js
@@ -216,6 +216,7 @@ storiesOf('Wallets|Send', module)
       currencyMaxIntegerDigits={11}
       currentNumberFormat={NUMBER_OPTIONS[0].value}
       validateAmount={promise(true)}
+      validateAssetAmount={promise(true)}
       calculateTransactionFee={promise(true)}
       walletAmount={new BigNumber(123)}
       assets={sendFormAssetData}
@@ -240,6 +241,7 @@ storiesOf('Wallets|Send', module)
       currencyMaxIntegerDigits={11}
       currentNumberFormat={NUMBER_OPTIONS[0].value}
       validateAmount={promise(true)}
+      validateAssetAmount={promise(true)}
       calculateTransactionFee={promise(true)}
       assets={sendFormAssetData}
       addressValidator={() => true}
@@ -264,6 +266,7 @@ storiesOf('Wallets|Send', module)
       currencyMaxIntegerDigits={11}
       currentNumberFormat={NUMBER_OPTIONS[0].value}
       validateAmount={promise(true)}
+      validateAssetAmount={promise(true)}
       calculateTransactionFee={promise(true)}
       assets={sendFormAssetData}
       addressValidator={() => true}
@@ -288,6 +291,7 @@ storiesOf('Wallets|Send', module)
       currencyMaxIntegerDigits={11}
       currentNumberFormat={NUMBER_OPTIONS[0].value}
       validateAmount={promise(true)}
+      validateAssetAmount={promise(true)}
       calculateTransactionFee={promise(true)}
       assets={sendFormAssetData}
       addressValidator={() => true}
@@ -312,6 +316,7 @@ storiesOf('Wallets|Send', module)
       currencyMaxIntegerDigits={11}
       currentNumberFormat={NUMBER_OPTIONS[0].value}
       validateAmount={promise(true)}
+      validateAssetAmount={promise(true)}
       calculateTransactionFee={promise({
         fee: new BigNumber(number('fee', 1)),
         minimumAda: new BigNumber(number('minimumAda', 1)),